@@ -33,12 +33,12 @@
     "promise": "8.0.1",
     "prop-types": "^15.6.0",
     "r": "0.0.5",
-    "react": "^16.0.0",
+    "react": "^16.8.6",
     "react-dev-utils": "^4.0.1",
-    "react-dom": "^16.0.0",
-    "react-redux": "^5.0.6",
-    "react-router-dom": "^4.2.2",
-    "redux": "^3.7.2",
+    "react-dom": "^16.8.6",
+    "react-redux": "^7.1.0",
+    "react-router-dom": "^5.0.1",
+    "redux": "^4.0.1",
     "redux-thunk": "^2.2.0",
     "style-loader": "0.18.2",
     "sw-precache-webpack-plugin": "0.11.4",
