@@ -1,7 +1,7 @@
 import axios from 'axios';
 
 const instance = axios.create({
-    baseURL: 'https://react-burger.firebaseio.com/'
+    baseURL: 'https://react-hooks-update.firebaseio.com'
 });
 
 export default instance;
