@@ -1,23 +1,26 @@
-import React, { Component } from 'react';
+import React, { useEffect } from 'react';
 import { Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
 
 import * as actions from '../../../store/actions/index';
 
-class Logout extends Component {
-    componentDidMount () {
-        this.props.onLogout();
-    }
+const logout = props => {
+  const { onLogout } = props;
 
-    render () {
-        return <Redirect to="/"/>;
-    }
-}
+  useEffect(() => {
+    onLogout();
+  }, [onLogout]);
 
-const mapDispatchToProps = dispatch => {
-    return {
-        onLogout: () => dispatch(actions.logout())
-    };
+  return <Redirect to="/" />;
 };
 
-export default connect(null, mapDispatchToProps)(Logout);
+const mapDispatchToProps = dispatch => {
+  return {
+    onLogout: () => dispatch(actions.logout())
+  };
+};
+
+export default connect(
+  null,
+  mapDispatchToProps
+)(logout);
