@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 
 import Order from '../../components/Order/Order';
@@ -7,42 +7,43 @@ import withErrorHandler from '../../hoc/withErrorHandler/withErrorHandler';
 import * as actions from '../../store/actions/index';
 import Spinner from '../../components/UI/Spinner/Spinner';
 
-class Orders extends Component {
-    componentDidMount () {
-        this.props.onFetchOrders(this.props.token, this.props.userId);
-    }
+const orders = props => {
+  const { onFetchOrders } = props;
 
-    render () {
-        let orders = <Spinner />;
-        if ( !this.props.loading ) {
-            orders = this.props.orders.map( order => (
-                <Order
-                    key={order.id}
-                    ingredients={order.ingredients}
-                    price={order.price} />
-            ) )
-        }
-        return (
-            <div>
-                {orders}
-            </div>
-        );
-    }
-}
+  useEffect(() => {
+    onFetchOrders(props.token, props.userId);
+  }, [onFetchOrders]);
+
+  let orders = <Spinner />;
+  if (!props.loading) {
+    orders = props.orders.map(order => (
+      <Order
+        key={order.id}
+        ingredients={order.ingredients}
+        price={order.price}
+      />
+    ));
+  }
+  return <div>{orders}</div>;
+};
 
 const mapStateToProps = state => {
-    return {
-        orders: state.order.orders,
-        loading: state.order.loading,
-        token: state.auth.token,
-        userId: state.auth.userId
-    };
+  return {
+    orders: state.order.orders,
+    loading: state.order.loading,
+    token: state.auth.token,
+    userId: state.auth.userId
+  };
 };
 
 const mapDispatchToProps = dispatch => {
-    return {
-        onFetchOrders: (token, userId) => dispatch( actions.fetchOrders(token, userId) )
-    };
+  return {
+    onFetchOrders: (token, userId) =>
+      dispatch(actions.fetchOrders(token, userId))
+  };
 };
 
-export default connect( mapStateToProps, mapDispatchToProps )( withErrorHandler( Orders, axios ) );
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(withErrorHandler(orders, axios));
